+++ demo/js/index.js
@@ -48,7 +48,6 @@ window.unbxdSearch = new UnbxdSearch({
   hashMode:true,
   updateUrls:true
 });
-
 window.unbxdSearch.updateConfig({
   searchBoxSelector: document.getElementById("unbxdInput"),
   searchTrigger: "click",
@@ -87,13 +86,14 @@ window.unbxdSearch.updateConfig({
   facet: {
     facetsEl: document.getElementById("facetsWrapper"),
     selectedFacetsEl: document.getElementById("selectedFacetWrapper"),
-    rangeFacetEl: document.getElementById("rangeFacetWrapper"),
-    multiLevelFacetEl: document.getElementById("bucketedFacetWrapper"),
     applyMultipleFilters:false,
     defaultOpen:"FIRST",
-    facetMultiSelect:false,
+    facetMultiSelect:true,
     onFacetLoad:function(facets){
-    }
+    },
+    isCollapsible:true,
+    isSearchable:true,
+    viewMore:true,
   },
   pagination: {
     type:'FIXED_PAGINATION',
@@ -142,6 +142,106 @@ window.unbxdSearch.updateConfig({
   },
   actionCallback:function(e, ctx) {
   },
-  onCallBack:unbxdCallbackEcma
-});
+  onEvent:unbxdCallbackEcma
+})
+
+window.unbxdSearch.updateConfig({
+  searchBoxSelector: document.getElementById("unbxdInput"),
+  searchTrigger: "click",
+  searchButtonSelector: document.getElementById("searchBtn"),
+  products:{
+    el:document.getElementById("searchResultsWrapper"),
+    productType:"SEARCH",
+    productClick: function(product,e) {
+      console.log(product,"product,index",e);
+    },
+    productAttributes: [
+      "title",
+      "uniqueId",
+      "price",
+      "sku",
+      "imageUrl",
+      "displayPrice",
+      "salePrice",
+      "sortPrice",
+      "productDescription",
+      "unbxd_color_mapping",
+      "colorName",
+      "color",
+      "size",
+      "listPrice",
+      "autosuggest"
+    ]
+  },
+  spellCheck: {
+    enabled: true,
+    el: document.getElementById("didYouMeanWrapper")
+  },
+  noResults: {
+      el: document.getElementById("noResultWrapper")
+  },
+  facet: {
+    facetsEl: document.getElementById("facetsWrapper"),
+    selectedFacetsEl: document.getElementById("selectedFacetWrapper"),
+    applyMultipleFilters:false,
+    defaultOpen:"FIRST",
+    facetMultiSelect:true,
+    onFacetLoad:function(facets){
+    },
+    isCollapsible:true,
+    isSearchable:true,
+    viewMore:true,
+  },
+  pagination: {
+    type:'FIXED_PAGINATION',
+    el:document.querySelector("#paginationContainer"),
+    onPaginate:function(data) {
+    }
+  },
+  breadcrumb: {
+    el: document.getElementById("breadcrumpContainer")
+  },
+  pagesize: {
+    el: document.getElementById("changeNoOfProducts")
+  },
+
+  sort: {
+    el: document.getElementById("sortWrapper"),
+    options: [
+      {
+        value: "sortPrice desc",
+        text: "Price High to Low"
+      },
+      {
+        value: "sortPrice asc",
+        text: " Price Low to High"
+      }
+    ]
+  },
+  loader :{
+      el:document.getElementById("loaderEl")
+  },
+  productView:{
+    el: document.getElementById("productViewTypeContainer"),
+    viewTypes: "GRID"
+  },
+  banner: {
+    el: document.getElementById("bannerContainer"),
+    count: 1
+  },
+  swatches: {
+    enabled: true,
+    attributesMap: {
+      swatchList: "color",
+      swatchImgs: "unbxd_color_mapping",
+      swatchColors: "color"
+    }
+  },
+  actionCallback:function(e, ctx) {
+  },
+  onEvent:unbxdCallbackEcma
+})
+
+
+
 //window.unbxdSearch.initialize();