+++ src/modules/facets/actions.js
@@ -5,12 +5,18 @@ const findChangedFacet = function(e) {
     const {
         facetName,
         facetAction,
-        id
+        id,
+        action,
+        end,
+        start
     } = dataSet;
     const qState = this.getStateFromUrl();
     const selectedfacets = this.getSelectedFacets();
     const ln = (selectedfacets) ?Object.keys(selectedfacets).length:0;
     const ql = Object.keys(qState.selectedFacets).length;
+    const {
+        productType
+    } = this.options
     if(facetAction === this.events.changeFacet) {
         const selectedfacetInfo = this.getSelectedFacet(facetName);
         const selectedOpt = {
@@ -19,7 +25,7 @@ const findChangedFacet = function(e) {
             facetData : selectedfacetInfo
         }
         this.viewState.lastAction = "addedAFacet";
-        this.options.onCallBack(this,this.events.facetClick, {
+        this.options.onEvent(this,this.events.facetClick, {
             facetName,
             facetData:selectedfacetInfo
         });
@@ -34,7 +40,7 @@ const findChangedFacet = function(e) {
         if(this.findSelectedFacet(facetName)) {
             this.viewState.lastAction = "deletedAfacet";
             this.deleteAFacet.bind(this)(facetName, id);
-            this.options.onCallBack(this,facetAction, {
+            this.options.onEvent(this,facetAction, {
                 facetName
             });
             this.getCallbackActions({
@@ -53,7 +59,7 @@ const findChangedFacet = function(e) {
         if(this.findSelectedFacet(facetName)) {
             this.viewState.lastAction = "deletedAfacet";
             this.deleteAFacet.bind(this)(facetName);
-            this.options.onCallBack(this,facetAction, {
+            this.options.onEvent(this,facetAction, {
                 facetName
             });
             this.getCallbackActions({
@@ -80,19 +86,30 @@ const findChangedFacet = function(e) {
         this.setPageStart(0);
         this.getResults();
     }
-    this.renderFacets();
-}
-const onClickRangeFacet = function(e) {
-    const {
-        action,
-        facetName,
-        end,
-        start
-    } = e.target.dataset;
-    const self = this;
-    const q = this.getStateFromUrl();
+
+    if(action === this.actions.setCategoryFilter) {
+        if(productType === "SEARCH") {
+            this.setCategoryFilter(dataSet);
+            this.options.onEvent(this,action, dataSet);
+        } else {
+            this.setCategoryId(dataSet, this);
+        }
+        this.getResults();
+        this.getCallbackActions(dataSet,'facetClick');
+    }
+    if(action === this.actions.clearCategoryFilter) {
+        if(productType === "SEARCH") {
+            this.deleteCategoryFilter(dataSet);
+            this.options.onEvent(this,action, dataSet);
+        } else {
+            this.setCategoryId(dataSet, this);
+        }
+        this.getResults();
+        this.getCallbackActions(dataSet,'facetClick');
+    }
     const ranges = this.state.rangeFacet[facetName];
-    const isSelections = Object.keys(q.rangeFacet);
+    const isSelections = Object.keys(qState.rangeFacet);
+    
     if(action === "setRange") {
         let already = false
         if(ranges &&ranges.length === 1 && isSelections.length > 0) {
@@ -130,9 +147,8 @@ const onClickRangeFacet = function(e) {
         }
     }
     if(action === this.actions.applyRange) {
-
         this.applyRangeFacet();
-        this.options.onCallBack(this,action, {
+        this.options.onEvent(this,action, {
             facetName
         });
         this.getCallbackActions({
@@ -142,42 +158,18 @@ const onClickRangeFacet = function(e) {
     if(action === this.actions.clearPriceRange && facetName) {
         this.clearARangeFacet(facetName);
         this.getResults.bind(this)();
-        this.options.onCallBack(this,action, {
+        this.options.onEvent(this,action, {
             facetName
         });
         this.getCallbackActions({
             facetName
         },'facetClick');
     }
+    this.viewState.lastAction = action;
+    //this.renderFacets();
 }
-const onBucketedFacet = function(e) {
-    const data = e.target.dataset;
-    const {
-        productType
-    } = this.options
-    if(data.action === this.actions.setCategoryFilter) {
-        if(productType === "SEARCH") {
-            this.setCategoryFilter(data);
-            this.options.onCallBack(this,data.action, data);
-        } else {
-            this.setCategoryId(data, this);
-        }
-        this.getResults();
-        this.getCallbackActions(data,'facetClick');
-    }
-    if(data.action === this.actions.clearCategoryFilter) {
-        if(productType === "SEARCH") {
-            this.deleteCategoryFilter(data);
-            this.options.onCallBack(this,data.action, data);
-        } else {
-            this.setCategoryId(data, this);
-        }
-        this.getResults();
-        this.getCallbackActions(data,'facetClick');
-    }
-}
+
+
 export {
-    findChangedFacet,
-    onClickRangeFacet,
-    onBucketedFacet
+    findChangedFacet
 }