+++ src/modules/facets/setFacetMethods.js
@@ -1,41 +1,51 @@
-import renderTextFacets from './index';
 import {
-    findChangedFacet,
-    onClickRangeFacet,
-    onBucketedFacet
+    findChangedFacet
 } from './actions';
-function renderRangeFacets() {
-    const rangeFacets = this.getRanges();
+
+function renderRangeFacet(facet,isExpanded) {
+    const {
+        facetName
+    } = facet;
+    const {
+        openFacet,
+        closeFacet
+    } = this.cssList;
+    const {
+        isCollapsible
+    } = this.options.facet;
     const selectedRanges = this.state.rangeFacet;
-    if(rangeFacets.length > 0) {
-        return this.options.facet.rangeTemplate.bind(this)(rangeFacets,selectedRanges);
-    } else {
-        return ``;
+    const facetUI =  this.options.facet.rangeTemplate.bind(this)([facet],selectedRanges);
+    let styles = (isExpanded)? openFacet:closeFacet;
+    if(!isCollapsible) {
+        styles="";
     }
+    return `<div id="${facetName}" class="UNX-facet-item-d range-facets-block ${styles}">${facetUI}<div>`
 }
+
 function setRangeFilter(data){
     this.setRangeFacet(data)
 }
-function renderBucketedUI() {
-    const bucketedFacets = this.getBucketedFacets();
+function renderMultiLevelFacet(bucketedFacet,isExpanded) {
     const breadCrumb = this.getBreadCrumbsList();
     let bucketedUi = ``;
-    const {
-        multiLevelFacetSelector
-    } = this.options.facet;
     const self = this;
-    bucketedFacets.forEach(facet => {
-        const {
-            displayName,
-            values,
-            facetName,
-        } = facet;
-        const isExpanded =  self.isExpandedFacet(facetName);
-        const valueUI = self.options.facet.multiLevelFacetTemplate.bind(this)(facet,breadCrumb);
-        bucketedUi += self.options.facet.facetTemplate.bind(self)(facet, valueUI,isExpanded,null);
-    });
-    return bucketedUi;
-    //return this.options.facet.multiLevelFacetTemplate.bind(this)(bucketedFacet,breadCrumb);
+    const {
+        openFacet,
+        closeFacet
+    } = this.cssList;
+    const{
+        facetName
+    } = bucketedFacet;
+    const {
+        isCollapsible
+    } = this.options.facet;
+    const valueUI = self.options.facet.multiLevelFacetTemplate.bind(this)(bucketedFacet,breadCrumb);
+    bucketedUi += self.options.facet.facetTemplate.bind(self)(bucketedFacet, valueUI,isExpanded,null);
+    let styles = (isExpanded)? openFacet:closeFacet;
+    if(!isCollapsible) {
+        styles="";
+    }
+    return `<div id="${facetName}" class="UNX-facet-item-d UNX-multilivel-facets-block UNX-multilevel-block ${styles}">${bucketedUi}<div>`
 }
 const isSelectedRange = function(facetName,range){
     const selections = this.getSelectedRanges(facetName);
@@ -53,7 +63,6 @@ const clearAllFacets = function() {
     this.state.selectedFacets = {};
     this.state.rangeFacet = [];
     this.state.categoryFilter = {};
-    this.renderRangeFacets();
 }
 const isExpandedFacet = function(facetName){
     const {
@@ -75,11 +84,21 @@ const getSearchFacetsText = function(facet) {
 
 const reRenderTextFacet = function(facetName) {
     const {
-        textFacetWrapper
+        textFacetWrapper,
+        facetClass
     } = this.options.facet;
-    const textFacet = this.getAFacetByName(facetName);
-    const textUi = this.renderTextFacets(textFacet, this.getSelectedFacets(), true);
-    document.querySelector(`#${facetName} .${textFacetWrapper}`).innerHTML = Array.isArray(textUi) ? textUi.join(""):textUi;
+    const facetSearchTxt = this.getSearchFacetsText(facetName) || "";
+    const items = document.querySelectorAll(`#${facetName} .${facetClass}`);
+    items.forEach(item => {
+        const {
+            id
+        } = item.dataset || {};
+        if(id && id.toUpperCase().indexOf(facetSearchTxt.toUpperCase()) >= 0) {
+            item.classList.remove("UNX-search-hidden")
+        } else {
+            item.classList.add("UNX-search-hidden")
+        }
+    })
 }
 const getAFacetByName = function(name) {
     const facets = this.getFacets();
@@ -94,23 +113,91 @@ const getAFacetByName = function(name) {
     }
     return selected;
 }
-const getAllFacets = function() {
+const sortFacets = (a,b) => {return (a.position > b.position) ? 1 : -1}
+const getAllTextFacets = function() {
     const text = this.getFacets();
-    const ranges = this.getRangeFacets();
-    const multiFacets = this.getBucketedFacets();
-    return [...multiFacets,...text,...ranges];
+    return text.map(item => {
+        item.facetType = "text";
+        return item;
+    })
 }
+const getAllRangeFacets = function() {
+    
+    const ranges = this.getRanges();
+    return ranges.map(item => {
+        item.facetType = "range";
+        return item;
+    })
+}
+const getAllCategory = function() {
+    const multiFacets = this.getBucketedFacets();
+    return multiFacets.map(item => {
+        item.facetType = "category";
+        return item;
+    })
+}
+
+const getAllFacets = function() {
+    const texts = this.getAllTextFacets();
+    const ranges = this.getAllRangeFacets();
+    const multiFacets = this.getAllCategory();
+    let all  = [...multiFacets,...texts,...ranges];
+    return all.sort(sortFacets);
+
+}
+const renderTextFacet = function(facetItem,selectedFacet,isExpanded,facetSearchTxt, onlyValues) {
+    let valuesUI = [];
+    const {
+        facetName,
+        values = []
+    } = facetItem;
+    const {
+        openFacet,
+        closeFacet
+    } = this.cssList;
+    let selectedFacetsUI = ``;
+    const self = this;
+    const {
+        isCollapsible
+    } = this.options.facet;
+    if(values.length > 0) {
+        valuesUI = values.map((value, index) => {
+            const  {
+                name,
+            } = value;
+            let selected = false;
+            if(selectedFacet) {
+                selected = selectedFacet.some((facet) => {
+                    return facet.name === name
+                })
+            }
+            if(selected) {
+                if(self.options.facet.selectedFacetsEl) {
+                    selectedFacetsUI += self.options.facet.selectedFacetTemplate.bind(self)(facetItem,value);
+                }
+                return self.options.facet.selectedFacetTemplate.bind(self)(facetItem,value,facetSearchTxt)
+            } else{
+                return self.options.facet.facetItemTemplate.bind(self)(facetItem, value,facetSearchTxt)
+            }
+        });
+    }
+    if(onlyValues) {
+        return valuesUI.join('');
+    }
+    const facetUI = this.options.facet.facetTemplate.bind(this)(facetItem, valuesUI.join(''),isExpanded,facetSearchTxt);
+    let styles = (isExpanded)? openFacet:closeFacet;
+    if(!isCollapsible) {
+        styles="";
+    }
+    return `<div id="${facetName}" class="UNX-facet-item-d ${styles}">${facetUI}<div>`
+}
+
 
 
 const setFacets = (prototype) => {
     prototype = Object.assign(prototype,{
-        renderTextFacets,
         findChangedFacet,
-        onClickRangeFacet,
-        renderRangeFacets,
         setRangeFilter,
-        renderBucketedUI,
-        onBucketedFacet,
         isSelectedRange,
         clearAllFacets,
         isExpandedFacet,
@@ -118,24 +205,31 @@ const setFacets = (prototype) => {
         getSearchFacetsText,
         reRenderTextFacet,
         getAFacetByName,
-        getAllFacets
+        getAllFacets,
+        getAllTextFacets,
+        getAllRangeFacets,
+        getAllCategory,
+        renderTextFacet,
+        renderRangeFacet,
+        renderMultiLevelFacet
     })
 };
 
 export {
     setFacets as default,
-    renderTextFacets,
     findChangedFacet,
-    onClickRangeFacet,
-    renderRangeFacets,
     setRangeFilter,
-    renderBucketedUI,
-    onBucketedFacet,
     clearAllFacets,
     isExpandedFacet,
     setSearchFacetsText,
     getSearchFacetsText,
     reRenderTextFacet,
     getAFacetByName,
-    getAllFacets
+    getAllFacets,
+    getAllTextFacets,
+    getAllRangeFacets,
+    getAllCategory,
+    renderTextFacet,
+    renderRangeFacet,
+    renderMultiLevelFacet
 };