+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
-##### 0.0.1
+## [0.1.8] - 2020-09-16
 
-- 🎉 First Commit!
+### Changed
+- facet configuration changes, only need to add the facetsEl under facet
+- removed rangeFacetEl and multiLevelFacteEl
+- onCallBack to on Event
+- no need of updateConfig as mandatory step
+
+
+### Added
+- configuration for view more functionality added, updated in read me
+- added animaion support for the facets collapse
+- added support for positioning the facets